Added function to skip tests if driver isn't available

+++ tests/Snok/MysqlCRUDTest.php
@@ -40,6 +40,7 @@ class MysqlCRUDTest extends \PHPUnit_Framework_TestCase {
 
     public static function tearDownAfterClass()
     {
+        if (!self::$dbh) return;
         self::$dbh->exec("DROP TABLE people");
         self::$dbh->exec("DROP TABLE species");
         self::$dbh->exec("DROP TABLE tools");
@@ -48,6 +49,9 @@ class MysqlCRUDTest extends \PHPUnit_Framework_TestCase {
     }
 
     private function setupEntity(&$instance) {
+        if (!self::$dbh) {
+            $this->markTestSkipped('The mysql database isn\'t available, check config and server.');
+        }
         $refObject = new \ReflectionClass($instance);
         $properties = $refObject->getProperties(\ReflectionProperty::IS_PROTECTED);
         foreach($properties as $property) {
